requirements.txt
# FastAPI & ASGI server
fastapi==0.110.0
uvicorn[standard]==0.29.0

# LangChain core
langchain==0.1.20
langchain-community==0.0.36
langchain-core==0.1.54
langchain-openai==0.1.6

# LLMs (you can swap out models/providers as needed)
openai==1.30.1

# SQL Support (MySQL via SQLAlchemy)
SQLAlchemy==2.0.30
pymysql==1.1.1

# MongoDB client
pymongo==4.7.2

# Environment variable management
python-dotenv==1.0.1

# Optional: Better JSON handling (optional, for convenience)
orjson==3.10.3

# Optional: Structured logging
loguru==0.7.2
